refactor: update error handling imports across handler files
