+++ crates/forge_server/src/lib.rs
@@ -5,7 +5,4 @@ mod routes;
 mod server;
 
 pub use config::ServerConfig;
-pub use error::{Error, Result};
-// Re-exports
-pub use forge_domain::AgentMessage;
 pub use server::ForgeServer;
